+++ tests/test_cli_installed.py
@@ -45,7 +45,7 @@ class TestInstalled(CLISubCommandTests, CLITableTests):
         )
 
     def test_json_parsing(self, invoke, subcmd):
-        result = invoke("--output-format", "json", subcmd)
+        result = invoke("--table-format", "json", subcmd)
         assert result.exit_code == 0
         data = json.loads(result.stdout)