+++ tests/test_cli_managers.py
@@ -65,7 +65,7 @@ class TestManagers(CLISubCommandTests, CLITableTests):
         self.check_manager_selection(result, set())
 
     def test_json_parsing(self, invoke, subcmd):
-        result = invoke("--output-format", "json", subcmd)
+        result = invoke("--table-format", "json", subcmd)
         assert result.exit_code == 0
         data = json.loads(result.stdout)